backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.2
DateTime==4.4
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.75.1
future==0.18.2
h11==0.13.0
idna==3.3
passlib==1.7.4
pyArango==2.0.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
zope.interface==5.4.0
